% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_functions.R
\name{roc_eval_text}
\alias{roc_eval_text}
\title{Generate Rd from text with evaluated inline code and code blocks}
\usage{
roc_eval_text(roclet, input)
}
\arguments{
\item{roclet}{Name of roclet to use for processing.}

\item{input}{Source string}
}
\value{
Same as \code{\link[roxygen2]{roc_proc_text}}.
}
\description{
\code{roc_eval_text} is an upgraded version of \code{\link[roxygen2]{roc_proc_text}}
that evaluates inline and block code before generating Rd.
}
\note{
Change log:
\itemize{
\item{0.1.0 Xiurui Zhu - Initiate the function.}
}
}
\examples{
fun_text <- '
#\' This is a `r nrow(iris)`-row matrix.
#\'
#\' This matrix has
#\' ```{r results="hold"}
#\' ncol(iris)
#\' ```
#\' columns.
my_fun <- function() iris
'
roc_eval_text(roxygen2::rd_roclet(), fun_text)[[1L]]
}
\author{
Xiurui Zhu
}
