% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-roclang.R
\docType{package}
\name{roclang-package}
\alias{roclang-package}
\alias{roclang}
\title{roclang: A package for diffusing function documentations into 'roxygen' comments}
\description{
The 'roclang' package facilitates efficient diffusing of content across function documentations.
Sections, parameters or dot parameters are extracted from function documentations and turned into valid Rd character strings,
which are ready to diffuse into the 'roxygen' comments of another function by inserting inline code.
}
\note{
Change log:
\itemize{
\item{0.1.1 Xiurui Zhu - Initiate the document.}
}
}
\section{Functions}{

\itemize{
\item Text extraction and manipulation function: \code{\link{extract_roc_text}}.
\item Rd evaluation and compilation function: \code{\link{roc_eval_text}}.
}
}

\author{
Xiurui Zhu
}
